+++ Antidote/FriendListCell.swift
@@ -64,7 +64,7 @@ class FriendListCell: BaseCell {
 
         let image = UIImage(named: "right-arrow")!.flippedToCorrectLayout()
         arrowImageView = UIImageView(image: image)
-        arrowImageView.setContentCompressionResistancePriority(UILayoutPriorityRequired, for: .horizontal)
+        arrowImageView.setContentCompressionResistancePriority(UILayoutPriority.required, for: .horizontal)
         contentView.addSubview(arrowImageView)
     }