+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Migrating to toxcore 0.2.2.
 - Migrating to Realm 3.1.0.
+- Migrating to Swift 4.
 
 ## [1.3.5] - 2017-08-23
 ### Changed